Handle case where SysUI, AAOS or Auto controller use Media3
All three util methods currently check if the controller is of
a legacy/platform type, which may not be true in all future cases.

Removing these checks allows to identify these controllers even
if they are implemented in Media3. The SysUI controller needs
some additional adjustments for all the cases where it's replaced
to and from the media notification controller. These translations
have to account for the fact the the original SysUI controller
may be part of the Media3 or platform controller list.

PiperOrigin-RevId: 780018652
(cherry picked from commit 10d7c65e5e773cea55fc554b0429a072434662f4)