Don't enforce available commands when connected to legacy session app
This change removes the check for available commands in `sendCustomCommand` of `MediaBrowserImplLegacy.java`. This is in parity with the behavior of legacy controllers/browsers when connected to a legacy app.

The test `sendCustomCommandWithMediaItem_commandButtonNotAvailable_permissionDenied` was updated to reflect this change.

PiperOrigin-RevId: 780580897
(cherry picked from commit 3174a48b37221c3b4658d4c78c69a0c27e9f6da1)