Deflake mute-unmute MediaController test
Waiting for the `unmuteLatch` alone was causing flakiness because the `eventsRef` would not be initialized, leading to NPE during assertions.

Changing the assertions to wait for the `eventLatch` instead exposed that there were never 2 volume_changed events, only 1, given a single `postAndSync` (unlike `unmute_twoControllers` test)

Instead, the 3 latches were combined into a single one and the event count expected as 1.

#cherrypick

PiperOrigin-RevId: 781067400