Do not show Snackbar message for automatically added items
When the main page becomes visible on the screen, 4 items are automatically added to the seqeuence(s). With each added item, we present a snackbar message to signal adding it.
However, the snackbar message doesn't matter much for these items, and we'd need to wait for them to be disappear before we click on preview.

PiperOrigin-RevId: 781091774