[ui-compose] Add ProgressStateWithTickCount to get position/duration
This ProgressState stateholder class is optimized for visual components that are limited in their UI-element representation (e.g. number of pixels in a circle/bar). The "tick" or "step" in such components is atomic due to some non-media nature (e.g. can't tick more often than a pixel). Hence, the polling of the current position from the Player can be bandlimited by how often visible progress will be noticeable to the user.

[demo-compose] Add HorizontalLinearProgressIndicator, aka timebar.

PiperOrigin-RevId: 781987240