Clean up some preload timeout logic values in ImaAdsLoader
We used two different preload timeouts, one hardcoded to 4 seconds
and one configurable (default 10 seconds) to detect slightly
different versions of the same problem that IMA doesn't preload
or fail an ad metadata load. This can be aligned to use the same
(configurable) timeout.

Secondly, this timeout should not be less than the timeout we
configured on the IMA SDK itself to load the VAST data. Otherwise
IMA has no chance of timing out internally first to potentially
provide a better experience (e.g. loading part of the ads instead
of marking the entire ad group as failed).

PiperOrigin-RevId: 783298825