Fix duration clipping when updating placeholder period != first
When we receive a new Timeline while the player is still using
placeholder data, we only re-resolve the originally requested
position if it's in the first period because the
period-window-offset may have changed. For all other periods,
this isn't needed because they start at position 0 by definition.

Calling Timeline.getPeriodPositionUs has another side effect
though to trim the position to be within the duration of the
period. This still needs to be done even for subsequent periods
in case they changed their duration from a placeholder to a
new value.

This can be nicely tested in ConcatenatingMediaSource2Test,
with a test case seeking beyond the overall duration. The
test setup already has a version where the duration gets
resolved later.

PiperOrigin-RevId: 783347448