Fix HDR sequence being accidentally tonemapped to SDR
`PlaybackVideoGraphWrapper` checks if the color transfer of a format is supported; if not, it tonemaps to SDR.

The color transfer check happens using `GlUtil.isExtensionSupported()` method, which retrieves the display by calling `EGL14.eglGetDisplay(EGL14.EGL_DEFAULT_DISPLAY)`. This call only retrieves the display without initializing it. When previewing a composition for the first time, the EGL display was not initialized when doing the check, so GlUtil.isExtensionSupported() returned false. This showed as if HDR is not supported on the device so it tonemapped to SDR.

The solution is to get the display by calling `GlUtil.getDefaultEglDisplay()` as it initializes the display before getting it.

PiperOrigin-RevId: 783369637