[HLS] Do not try to seek inside buffer when no chunks are in the buffer
When a `HlsSampleStreamWrapper` has no tracks enabled after track selection, it will clear the `mediaChunks` as well as the the sample queue. However, when the tracks of it are newly re-enabled, we should not attempt to seek inside buffer while there is still no chunks at least partially in the buffer.

Issue: androidx/media#2598

#cherrypick

PiperOrigin-RevId: 783754042