Add max time diff setting for auto-advancing FakeClock
The auto advancing feature assumes there is a regular driver that
pushes the time forward in small-ish increments (as done in ExoPlayer).

This logic can break if:
 - There is process generating new messages out of control of the
   clock (e.g. a Loader in ExoPlayer)
 - The message queue has another timed message much further in
   the future that was never meant to drive the clock forward.

This issue can be avoided by imposing a maximum expected time diff
for the clock auto-advancing mechanism. Choosing 1 second by default
should work well for all (Exo)Player use cases even with dynamic
scheduling of updates. Also adding a Builder to make this configurable
if required.

PiperOrigin-RevId: 784194600