Add `AnalyticsListener.onDroppedSeeksWhileScrubbing`
This gives a metric for how many seeks were dropped while the user was
scrubbing around, which is a proxy for how jerky or smooth their
scrubbing experience was.

The callback is fired when exiting scrubbing mode with a count of the
number of seeks dropped during that scrubbing 'session'. It is
deliberately documented more loosely (and similarly to the dropped
frames callback), to allow us to start calling it **during** scrubbing
mode in future if needed.

PiperOrigin-RevId: 785396481