Ensure RobolectricUtil.runLooperUntil doesn't advance time too far
The Robolectric implementation may accidentally advance time internally
if the only available message is in the future. This was never an
intended behavior of this util method, so adding a check to only execute
new messages within a reasonable time diff. The default of 1 second
matches the one used by FakeClock for the same purpose.

Also make this configurable on the one most used runMainLooperUntil
method and set the value in some of the ExoPlayer tests relying on
longer delays.

PiperOrigin-RevId: 785427931