Fix a `ParsableByteArray` limit bug in `FlacFrameReader`
Also clean-up the implementation a bit to avoid allocating two arrays
and copying between them. Instead we just allocate a single (large
enough) array, which also makes the code more concise.

Writing a test for this required `ParsableByteArray` to enforce its
limit, so I added that logic too behind a test-only (for now) gate. We
can use this gate in other tests as part of flushing out more limit
bugs as we work towards b/147657250.

PiperOrigin-RevId: 785501853