Try to officially disconnect controller in more cases
In cases where we intentionally ignore a connection request, we
don't currently attempt to call onDisconnected on the controller
in all cases. This means such controllers may not be aware they
are ignored. Even if some cases only happen with malformed or
malicious connection attempts, it's still useful to try and
disconnect the connection officially.

PiperOrigin-RevId: 785814501