Add detection for player stuck in BUFFERING state
This detection needs to happen on the application thread to capture
cases where the playback thread is not responsive (e.g. stuck in a
MediaCodec call).

As this is only a last resort fallback timeout, we choose a really
large value of 10 minutes at which we should be able to see some
progress even for data sources with long timeouts and retries. As
some apps may intentionally choose to wait for longer, this value
is configurable on ExoPlayer.

PiperOrigin-RevId: 785864050