Unflake MediaControllerListenerTest and MediaBrowserListenerTest
Addressing flakiness of `recursiveChangesFromListeners_reportConsistentValuesForAllListeners` and
`onEvents_whenNewCommandIsCalledInsideOnEvents_isCalledFromNewLooperIterationSet`

The test did not have any way to avoid that events are added to the list
to assert after the count down latch was counted down. This created either
failing asserts or a `ConcurrentModificationException`. Making sure to not
produce further events or to not add further events to the data structure
that is later used for asserting fixes the problem.

PiperOrigin-RevId: 786257610