Make some tests less prone to flakiness
Some playback tests may get caught in timeout conditions for
stuck playback during the initial phase of waiting for playback
to become ready. For real files in particular, this phase may
take an arbitrary amount of time due to the free-running loading
thread. This can be avoided by explicitly waiting for playback
to be ready before attempting to play.

PiperOrigin-RevId: 788396993