Update audio renderer isReady logic
We currently wait until either the upstream source is ready or the
audio sink has any pending data. This causes the renderer to become
ready too early (once data is available for decoding but before this
data has been pushed to the AudioSink). This can be checked by reducing
the condition to just checking the audio sink for pending data. Unlike
video, it's okay to not have upstream data available because the audio
sink maintains an internal buffer (usually part of the AudioTrack).

One ExoPlayerTest accidentally relied on this bug in its test setup that
really only wanted to wait for a single processed buffer.

PiperOrigin-RevId: 788893850