Forecast whether a frame will be dropped after effects
Use VideoFrameReleaseEarlyTimeForecaster to determine whether a frame is
likely to be late. The lateness is forecast before applying effects in
`InputVideoSink#handleInputFrame`. If a frame will likely be late, it is
dropped before effects are applied.

The forecaster is also updated with the actual early time for output frames.
`onVideoFrameProcessed` is called when a frame is about to be scheduled
for display, in `VideoFrameRenderControl`.

Pre-effect frame dropping can be controlled via
CompositionPlayer.Builder.experimentalSetLateThresholdToDropInputUs

PiperOrigin-RevId: 789232704