Make DownloadService ready for API 35
The service may be stopped by the system after running
for 6 hours without the app being in the foreground.

In such a case the system calls `onTimeout(int, int)` to
give the service the chance to gracefully terminate instead
of being suddenly crashed by the system a few instances later.

https://developer.android.com/about/versions/15/behavior-changes-15#datasync-timeout
Issue: androidx/media#2614
PiperOrigin-RevId: 789247191