Remove uncontrolled time advancement from ExoPlayerAssetLoaderTest
The test is intentionally non-deterministic (see comments about using
default clock and not a FakeClock). The randomness is introduced by
various sleep calls in the execution path, but while this is
happening the main test thread runs an infinite loop that keeps
increasing the 'current' time in an insane speed. This causes issues
when introducing timeouts on the main thread that get easily triggered
by this setup.

The issue can be fixed by also sleeping the main test thread a bit
before re-checking the finished condition and advancing the time.

PiperOrigin-RevId: 789731760