Prevent crash in `CmcdData` for non-hierarchical URIs
`CmcdData.removeFromUri` and `CmcdData.removeFromDataSpec` crashes with an `UnsupportedOperationException` if passed a non-hierarchical URI, such as a `data:` URI used for a DASH manifest. This occurred because the methods called `getQueryParameter` without first checking if the URI was hierarchical.

This change fixes the crash by adding a check for `uri.isHierarchical()` before any attempt to access query parameters.

Additionally, this change also refactors `CmcdData.removeFromDataSpec` to remove a redundant check.

Fixes: Issue: androidx/media#2688

#cherrypick

PiperOrigin-RevId: 789746287