Remove incorrect `checkNotNull` calls in `RepeatedTranscodeTest`
The value passed here is a primitive `long`, it can never be null.
This was revealed by a compilation failure as part of migrating to
Guava's `Preconditions` methods.

The explicit null check isn't needed here in general, because if
something earlier in the de-referencing chain is null the test will
fail with an NPE anyway (which is all that `checkNotNull` would do).

PiperOrigin-RevId: 791217138