Report AudioGraphInputAudioSink ended when it's not processed data
This happens when clipping ending position + seeking to the end position

This is visible when using SilenceMediaSource - and could happen for
progressive.

The `SilenceSampleStream`, after seeking to the end position, could output a
buffer that contains data past the end position. `ClippingSampleStream` on
reading the buffer, as the buffer contains more data past the end position,
sends EOS to the renderer. When the renderer sees the EOS,
`MCAR.renderToEndOfStream()` is called, which subsequently calls
`audioSink.playToEndOfStream()`. Now, if the `AudioGraphInputAudioSink` is
just created, its `outputGraphInput` has not been created yet, and will fail
when handling `playToEndOfStream()`

PiperOrigin-RevId: 792192320