Handle malformed `free` boxes in MP4 and fMP4
Add a workaround in the MP4 (fMP4) `Sniffer` and `Extractor` to handle `free` boxes with a size smaller than their header size. Such boxes are treated as having a size equal to the header, preventing sniff and parsing failures on files with this specific type of corruption.

Test file was created from the existing `sample.mp4` and modifying the size of `free` box.

Fixes: Issue: androidx/media#2567

#cherrypick

PiperOrigin-RevId: 794963590