Improve TestPlayer's handling of available commands
`Util.getAvailableCommands` is used by ExoPlayer and is an opinionated way to add and remove commands depending on the nature of media (ad, seekable, live etc) and location in the playlist. TestPlayer can take it as inspiration to provide a realistic scenario, making testing more streamlined. For example, asserting that the NextButton is disabled on the last item of the playlist, without having to explicitly remove the relevant commands after every operation.

PiperOrigin-RevId: 799561731