Open TestPlayer for modification
Subclasses of TestPlayer will be able to write to the inner state variable and customize handle* methods. The `state` variable will remain private and hence not need to be renamed (to avoid clashing with the synthetically generated `getState` method by Kotlin)

Since there is no way to organically move SimpleBasePlayer from STATE_BUFFERING to STATE_READY, we always end up with a buffering player after seeking and preparing. Even when playWhenReady is true, isPlaying remains false because the Player is not in STATE_READY. This change adds a customization to skip buffering, since there are no underlying resources being waited on anyway.

PiperOrigin-RevId: 799564850