Send custom commands to the session if not targeted to the service
When connected to a legacy session, custom commands are sent to the session
unless it is specifically targeted to the `MediaBrowserService` by using
a `ProgressListener` or when a custom action is contained in the set of
available media item command buttons as defined by the session.

Also completely ignore available commands when sending custom commands to
a legacy session.

PiperOrigin-RevId: 800009129