[ui-compose] Add command checks in button state onClick's
Compose UI prevents a human user from clicking on a disabled element by not invoking the `onClick` callback when `enabled` parameter is `false`. The element often appears differently for accessibility services (greyed out, hollow etc.)

However, media3 UI state holder classes allow a programmatic call to `onClick` that can be clashing with the state of the Player. The behaviour of calling a Player methods where the relevant command(s) is/are not available is not defined: a Player implementation is free to do anything: ignore the operation, throw an exception, do the operation regardless.

Between the options of "silently ignore" or "fail with exception", we are going with the latter to make the state holders more robust for usages in contexts that might not have the same guarantee as a Composables.

PiperOrigin-RevId: 800054825