Replace test uses of `SparseBooleanArray` with `boolean[]`
This is significantly faster than `SparseBooleanArray` (at a trade-off
of slightly increased memory consumption, but in a robolectric test
this isn't a significant concern).

The partial read arrays need to be able to grow to arbitrary lengths,
but the read error arrays only have to be one larger than the
underlying data.

PiperOrigin-RevId: 800847593