Create a new `media3-ui-compose-material3` module
Add NextButton and PreviousButton Composables to it.

-------

This module will be used for UI components built with Jetpack Compose using Material3 theming:

`implementation("androidx.media3:media3-ui-compose-material3:1.X.Y")`

The components from the compose demo app will be housed in the new module instead of `media3-ui-compose` which will remain free of `material3` depedency. That module provides more foundational components (like a Surface or higher-level Composables with Slots) that don't require UI decisions and as much customization. This module, on the other hand, has very opinionated UI, although we provide as much of a customisable API surface as possible.

PiperOrigin-RevId: 800950767