`TestSimpleBasePlayer`: Transition from `BUFFERING` to `READY`
This makes the player behave more like a realistic `Player`
implementation, which transitions into `BUFFERING` when `prepare()`
or a seek operation is called, and then transitions to `READY`
automatically after a delay. The delay defaults to 100ms, but is
customizable in a constructor parameter.

Previous to this change, a `prepare()` or seek operation would make
this player permanently enter `BUFFERING` state.

This removes the two cases in the codebase that needed to subclass
`TestSimpleBasePlayer` in order to make a seek transition directly to
`READY` instead of `BUFFERING`, and so it makes `TestSimpleBasePlayer`
final which will make it easier to evolve the API in future.

PiperOrigin-RevId: 801756056