Speed up untilPositionAtLeast
The method is currently using the background condition check
which sends a lot of messages to check the condition. When
waiting for a slightly longer playback, this considerably slows
down the test. As we are talking about a playback position,
we can make sure to trigger a main thread message just after
reaching the position to trigger the condition.

PiperOrigin-RevId: 801776840