Add timeout for stuck playing but not ending
This detects issues where the player is stuck in a state with
an advancing clock (e.g. stand-alone clock or interpolated
clock). We can only detect something is wrong once the playback
position exceeds the declared duration by an unreasonable amount
of time. The default is 1 minute to account for some expected
cases where the position exceeds the duration (e.g. inaccurate
MP3 position estimation or inaccurate duration metadata in file)

PiperOrigin-RevId: 801849124