Improve backwards compatiblity of new local playback defaults
If an existing app uses only one of the generic setters, they assumed
that the other setting is at its (streaming) default. To not violate
this assumption and create unexpected combinations of settings, we
should enforce this is still true. Only if no customization is made
or the new specific setters are used do the new defaults take effect.

PiperOrigin-RevId: 803388168