Add stuck suppressed detection
The player may become stuck if it has a suppression reason
(which is by definition temporary) and the suppression doesn't
resolve itself. Choosing a long timeout of 10 minutes to not
interfere with any valid suppressions.

Transient audio focus loss is always excluded as it's tracked
and handled by the audio system, and because it's the only
reason that has a known use case for a longer suppression (e.g.
for a phone call).

PiperOrigin-RevId: 803470983