Fix video get stuck in V+A -> A sequence
The issue is, track selection happens right after a period is prepared, which
could happen soon after loading the first V+A period because of pre-warming.
In this case the player would wrongly inform `PlaybackVideoGraphWrapper` that
the number of video sequence is zero, because of the track selection for the
audio only item. After this happens, the InputVideoSink won't receive any
input frames because of the checks in `shouldRenderToInputVideoSink()`

PiperOrigin-RevId: 804555402