[demo-compose] Refactor Compose demo
Instead of modifying MainActivity, we should have better encapsulation and separation of concerns. MainActivity should be the entry point with a single Composable that defines a screen. In the demo, it happens to be a fullscreen player and not much other functionality (other than a button that controls the aspect ratio)

In the future, if we want to add a playlist-choosing Screen which then navigates to the MediaPlayer screen, it will be easier to refactor code.

PiperOrigin-RevId: 805332120